+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 5.2.2 (2020-10-19)
+
+### Bug fixes
+
+* Avoid crash when initializing bugsnag in attachBaseContext
+  [#953](https://github.com/bugsnag/bugsnag-android/pull/953)
+
+* Prevent ConcurrentModificationException when setting redactedKeys
+  [#947](https://github.com/bugsnag/bugsnag-android/pull/947)
+
 ## 5.2.1 (2020-10-01)
 
 ### Bug fixes